# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sunrise Staff Housing Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buildings:
    get:
      operationId: listBuildings
      tags: [buildings]
      summary: List buildings
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of buildings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Building"
    post:
      operationId: createBuilding
      tags: [buildings]
      summary: Create building
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBuildingBody"
      responses:
        "201":
          description: Building created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
  /buildings/{id}:
    patch:
      operationId: updateBuilding
      tags: [buildings]
      summary: Update building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBuildingBody"
      responses:
        "200":
          description: Updated building
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
    delete:
      operationId: deleteBuilding
      tags: [buildings]
      summary: Delete building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # FLOORS
...
components:
  schemas:
...
    Building:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        status:
          type: string
      required:
        - id
        - propertyId
        - name
        - location
        - capacity
        - status
    CreateBuildingBody:
      type: object
      properties:
        propertyId:
          type: integer
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        status:
          type: string
      required:
        - propertyId
        - name
        - location
        - capacity
    UpdateBuildingBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        status:
          type: string
...